% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_RR_estimate.R
\name{adjust_RR_estimate}
\alias{adjust_RR_estimate}
\title{Adjustment formula}
\usage{
adjust_RR_estimate(R, pL, rV, f, nV, N)
}
\arguments{
\item{N}{}
}
\value{

}
\description{
Adjustment formula
}
