% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_graph.R
\name{adjustment_graph}
\alias{adjustment_graph}
\title{Plot adjusted relative risk estimates for varying p_L and r_V}
\usage{
adjustment_graph(
  R,
  N,
  nV,
  f,
  n_points = 201,
  label_x = 1.05,
  lwd = 1,
  margin_offset = 0.02
)
}
\arguments{
\item{margin_offset}{}
}
\value{

}
\description{
Plot adjusted relative risk estimates for varying p_L and r_V
}
