% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data at an individual level}
\usage{
sim_data(
  N = 11 * 10^6,
  `p(V)` = 0.4,
  `p(E|!V)` = 0.0014,
  R = 0.25,
  `p(E|V)` = `p(E|!V)` * R,
  `p(V*|V)` = 0.75,
  `p(V*|!V)` = 0,
  `p(E*|E)` = 0.75,
  `p(E*|!E)` = 0,
  `p(L|V*,E*)` = 0.75
)
}
\arguments{
\item{`p(L|V*, E*)`}{}
}
\description{
Simulate data at an individual level
}
