% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_binom.R
\name{sim_data_binom}
\alias{sim_data_binom}
\title{Simulate data at a population level}
\usage{
sim_data_binom(
  N = 11 * 10^6,
  `p(V)` = 0.4,
  `p(E|!V)` = 0.0014,
  RR = 0.25,
  `p(E|V)` = `p(E|!V)` * RR,
  `p(V*|V)` = 0.75,
  `p(E*|E)` = 0.75,
  `p(L|V*E*)` = 0.75,
  n_sims = 1000
)
}
\arguments{
\item{N}{The population size}

\item{n_sims}{}
}
\value{

}
\description{
Simulate data at a population level
}
