% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_multiplier.R
\name{reporting_multiplier}
\alias{reporting_multiplier}
\title{Title}
\usage{
reporting_multiplier(
  data = tibble(`p(V)`, `p(E|!V)`, `p(E|V)`, `p(V*|V)`, `p(V*)`, `p(E*|E)`,
    `p(L|V*E*)`, R),
  `p(V)` = 0.4,
  `p(E|!V)` = 0.0014,
  R = 0.25,
  `p(E|V)` = `p(E|!V)` * R,
  `p(V*|V)` = 0.75,
  `p(V*)` = `p(V*|V)` * `p(V)`,
  `p(E*|E)` = 0.75,
  `p(L|V*E*)` = 0.75
)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Title
}
