% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_multiplier.R
\name{reporting_multiplier}
\alias{reporting_multiplier}
\title{Title}
\usage{
reporting_multiplier(
  `p(V)` = 0.4,
  R = 0.25,
  `p(V*|V)` = 0.75,
  `p(L|V*E*)` = 0.75,
  N = 11 * 10^6,
  `N*` = N,
  f = (`N*` - N)/N,
  ...
)
}
\arguments{
\item{...}{does nothing}
}
\value{

}
\description{
Title
}
